Split DBManager into DbConnection + per-feature repositories
DBManager (1221 LOC) is replaced by:
  - DbConnection — owns the sqlite3 handle, shared mutex, executeQuery
  - schema.cpp — initializeSchema() runs CREATE TABLE + column migrations
  - MealsRepository — meals, ingredients, default-seed data
  - WorkoutsRepository — workouts, workout templates
  - GoogleTokensRepository — encrypted OAuth token storage

Each repository takes a shared_ptr<DbConnection>, so they share one
SQLite handle and serialize via its mutex.

Wiring updated:
  - MealFactory now takes MealsRepository (was DBManager)
  - GoogleOAuth now takes GoogleTokensRepository (was DBManager)
  - setupRoutes() takes meals + workouts repos directly; per-feature
    route registrars only see the repo they actually use
  - main.cpp constructs DbConnection once and hands each repo to the
    feature that needs it

Tests now construct only the repository they exercise. The combined
test_db_manager.cpp is split into test_meals_repository.cpp and
test_google_tokens_repository.cpp. 101 tests still green.
