+++ src/core/http/api_routes.cpp
@@ -5,12 +5,13 @@
 #include "features/workouts/workouts_routes.h"
 #include "integrations/google/google_routes.h"
 
-void setupRoutes(crow::App<RequestTimerMiddleware>& app, std::shared_ptr<DBManager> dbManager,
-                 MealFactory& factory, const Config& /*config*/,
+void setupRoutes(crow::App<RequestTimerMiddleware>& app,
+                 std::shared_ptr<MealsRepository> meals,
+                 std::shared_ptr<WorkoutsRepository> workouts, MealFactory& factory,
                  const std::shared_ptr<GoogleOAuth>& googleOAuth,
                  const std::shared_ptr<CalendarService>& calendarService) {
-    registerMealsRoutes(app, dbManager, factory);
-    registerWorkoutsRoutes(app, dbManager);
+    registerMealsRoutes(app, std::move(meals), factory);
+    registerWorkoutsRoutes(app, std::move(workouts));
     registerGoogleRoutes(app, googleOAuth, calendarService);
     registerStaticRoutes(app);
 }
