+++ src/core/http/api_routes.h
@@ -4,10 +4,10 @@
 
 #include <memory>
 
-#include "core/config/config_parser.h"
-#include "core/db/db_manager.h"
 #include "core/http/middleware.h"
 #include "features/meals/meal_factory.h"
+#include "features/meals/meals_repository.h"
+#include "features/workouts/workouts_repository.h"
 #include "integrations/google/calendar_service.h"
 #include "integrations/google/google_oauth.h"
 
@@ -18,7 +18,8 @@
  * static files). New feature modules should add their own
  * `register<Feature>Routes()` call here.
  */
-void setupRoutes(crow::App<RequestTimerMiddleware>& app, std::shared_ptr<DBManager> dbManager,
-                 MealFactory& factory, const Config& config,
+void setupRoutes(crow::App<RequestTimerMiddleware>& app,
+                 std::shared_ptr<MealsRepository> meals,
+                 std::shared_ptr<WorkoutsRepository> workouts, MealFactory& factory,
                  const std::shared_ptr<GoogleOAuth>& googleOAuth,
                  const std::shared_ptr<CalendarService>& calendarService);
