+++ tests/features/meals/test_meal_factory.cpp
@@ -4,86 +4,68 @@
 #include <string>
 #include <tuple>
 
-#include "core/db/db_manager.h"
+#include "core/db/db_connection.h"
+#include "core/db/schema.h"
 #include "features/meals/meal.h"
 #include "features/meals/meal_factory.h"
+#include "features/meals/meals_repository.h"
 
 class MealFactoryTest : public ::testing::Test {
    protected:
     void SetUp() override {
-        auto dbManager = std::make_shared<DBManager>(":memory:");
-        dbManager->initializeSchema();
-        dbManager->seedDefaultMeals();
-        factory = std::make_unique<MealFactory>(dbManager);
+        auto conn = std::make_shared<DbConnection>(":memory:");
+        initializeSchema(*conn);
+        auto repo = std::make_shared<MealsRepository>(conn);
+        repo->seedDefaultMeals();
+        factory = std::make_unique<MealFactory>(repo);
     }
-    void TearDown() override {}
 
     std::unique_ptr<MealFactory> factory;
 };
 
-// Test creating turkey-burgers
 TEST_F(MealFactoryTest, CreateTurkeyBurgers) {
     auto meal = factory->createMeal("turkey-burgers");
-
     ASSERT_NE(meal, nullptr);
-    // Note: By default the DB preserves the "turkey-burgers" key name as the
-    // inserted Name The previous tests expected "Turkey Burgers", we need to
-    // check the actual value
     EXPECT_EQ(meal->getName(), "turkey-burgers");
 }
 
-// Test creating unknown meal
 TEST_F(MealFactoryTest, CreateUnknownMeal) {
     auto meal = factory->createMeal("unknown-meal");
-
     EXPECT_EQ(meal, nullptr);
 }
 
-// Test case sensitivity (DB searches usually are case sensitive unless COLLATE
-// NOCASE is used)
 TEST_F(MealFactoryTest, CaseSensitivity) {
-    auto meal1 = factory->createMeal("Turkey-Burgers");  // Wrong case
-    auto meal2 = factory->createMeal("TURKEY-BURGERS");  // Wrong case
-
-    EXPECT_EQ(meal1, nullptr);
-    EXPECT_EQ(meal2, nullptr);
+    EXPECT_EQ(factory->createMeal("Turkey-Burgers"), nullptr);
+    EXPECT_EQ(factory->createMeal("TURKEY-BURGERS"), nullptr);
 }
 
-// Test getAvailableMeals
 TEST_F(MealFactoryTest, GetAvailableMeals) {
     std::vector<std::tuple<int, std::string, std::string>> meals;
     factory->getAvailableMeals(meals);
-
     EXPECT_GT(meals.size(), 0);
 
-    // Check that expected meals are in the list
     bool hasTurkeyBurgers = false;
     bool hasChickenStirFry = false;
-
     for (const auto &mealTuple : meals) {
-        if (std::get<1>(mealTuple) == "turkey-burgers") {
-            hasTurkeyBurgers = true;
-        }
-        if (std::get<1>(mealTuple) == "chicken-stir-fry") {
-            hasChickenStirFry = true;
-        }
+        if (std::get<1>(mealTuple) == "turkey-burgers") hasTurkeyBurgers = true;
+        if (std::get<1>(mealTuple) == "chicken-stir-fry") hasChickenStirFry = true;
     }
-
     EXPECT_TRUE(hasTurkeyBurgers);
     EXPECT_TRUE(hasChickenStirFry);
 }
 
-// createMeal without an add-on set excludes all optional ingredients
 TEST_F(MealFactoryTest, CreateMealExcludesOptionalByDefault) {
-    auto dbManager = std::make_shared<DBManager>(":memory:");
-    dbManager->initializeSchema();
+    auto conn = std::make_shared<DbConnection>(":memory:");
+    initializeSchema(*conn);
+    auto repo = std::make_shared<MealsRepository>(conn);
+
     std::vector<Ingredient> ings = {
         Ingredient("Eggs", Measurement(2.0, MeasurementUnit::WHOLE), "None", false),
         Ingredient("Cottage Cheese", Measurement(1.0, MeasurementUnit::CUP), "None", true),
         Ingredient("Pumpkin", Measurement(0.25, MeasurementUnit::CUP), "None", true),
     };
-    dbManager->addMeal(Meal("pancakes-test", ings, "Breakfast"));
-    MealFactory localFactory(dbManager);
+    repo->addMeal(Meal("pancakes-test", ings, "Breakfast"));
+    MealFactory localFactory(repo);
 
     auto meal = localFactory.createMeal("pancakes-test");
     ASSERT_NE(meal, nullptr);
@@ -91,18 +73,19 @@ TEST_F(MealFactoryTest, CreateMealExcludesOptionalByDefault) {
     EXPECT_EQ(meal->getIngredients()[0].getName(), "Eggs");
 }
 
-// createMeal with a set of enabled add-ons includes only the matching optionals
 TEST_F(MealFactoryTest, CreateMealIncludesSelectedAddOns) {
-    auto dbManager = std::make_shared<DBManager>(":memory:");
-    dbManager->initializeSchema();
+    auto conn = std::make_shared<DbConnection>(":memory:");
+    initializeSchema(*conn);
+    auto repo = std::make_shared<MealsRepository>(conn);
+
     std::vector<Ingredient> ings = {
         Ingredient("Eggs", Measurement(2.0, MeasurementUnit::WHOLE), "None", false),
         Ingredient("Cottage Cheese", Measurement(1.0, MeasurementUnit::CUP), "None", true),
         Ingredient("Pumpkin", Measurement(0.25, MeasurementUnit::CUP), "None", true),
         Ingredient("Blueberry", Measurement(0.5, MeasurementUnit::CUP), "None", true),
     };
-    dbManager->addMeal(Meal("pancakes-test", ings, "Breakfast"));
-    MealFactory localFactory(dbManager);
+    repo->addMeal(Meal("pancakes-test", ings, "Breakfast"));
+    MealFactory localFactory(repo);
 
     auto meal = localFactory.createMeal("pancakes-test", {"Blueberry", "Pumpkin"});
     ASSERT_NE(meal, nullptr);
@@ -116,7 +99,6 @@ TEST_F(MealFactoryTest, CreateMealIncludesSelectedAddOns) {
     EXPECT_EQ(names.count("Cottage Cheese"), 0u);
 }
 
-// Test that all available meals can be creatable
 TEST_F(MealFactoryTest, AllAvailableMealsAreCreatable) {
     std::vector<std::tuple<int, std::string, std::string>> meals;
     factory->getAvailableMeals(meals);
